Refactor serialization logic: Moved `ensure_serializable` function to a new `utils.py` file for better organization and reusability. Updated `langchain_tool_adapter.py` to import this function, streamlining the code structure.
